FinalTest

+++ Personalportfolio/Controllers/AwardController.cs
@@ -1,8 +1,11 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using Personalportfolio.Data;
 using Personalportfolio.Data.Service;
 using Personalportfolio.Models;
+using Personalportfolio.Models.ViewModel;
+using System.Security.Claims;
 
 namespace Personalportfolio.Controllers
 {
@@ -10,34 +13,57 @@ namespace Personalportfolio.Controllers
     public class AwardController : Controller
     {
         private readonly IAwardService _service;
+        
+         
         public AwardController( IAwardService service)
         {
+            
             _service = service;
         }
+       
         public IActionResult Index()
         {
-            IEnumerable<Awards> data = _service.GetAll();
+
+            var userId = User.FindFirstValue("userId");
+            IEnumerable<Awards> data = _service.GetAll().Where(x=>x.UId.ToString() == userId).ToList();
             return View(data);
         }
+
+
+        [HttpGet]
+        public IActionResult FinalTest()
+        {
+
+            var userId = User.FindFirstValue("userId");
+            IEnumerable<Awards> data = _service.GetAll().Where(x => x.UId.ToString() == userId);
+            return View(data);
+        }
+
+
         [HttpGet]
         public IActionResult Create()
         {
-           
+            var usermail = User.FindFirstValue("userId");
+            ViewBag.email = usermail;
             return View();
         }
 
         [HttpPost]
         public IActionResult Create(Awards awards)
         {
-            try
-            {
-                _service.Add(awards);
-                return RedirectToAction("Index");
 
-            }
-            catch (Exception ex)
+            if (ModelState.IsValid)
             {
-                Console.WriteLine(ex.Message);
+
+                try
+                {
+                    _service.Add(awards);
+                    return RedirectToAction("Index");
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                }
             }
             return View();
         }
@@ -45,15 +71,29 @@ namespace Personalportfolio.Controllers
         [HttpGet]
         public IActionResult Edit(int id)
         {
+            var usermail = User.FindFirstValue("userId");
             Awards data = _service.GetById(id);
-            return View(data);
+              if(data.UId.ToString() == usermail)
+            {
+                return View(data);
+            }
+            return View();
+           
         }
 
         [HttpPost]
         public IActionResult Edit(Awards awards)
         {
-            _service.Update(awards);
-            return RedirectToAction("Index");
+            if(ModelState.IsValid)
+            {
+                _service.Update(awards);
+                return RedirectToAction("Index");
+            }
+            else
+            {
+                return View();
+            }
+           
         }
 
         public async Task<IActionResult> Delete(int id)
@@ -65,8 +105,17 @@ namespace Personalportfolio.Controllers
         [HttpPost, ActionName("Delete")]
         public IActionResult Deleted(int id)
         {
-            _service.Delete(id);
-            return RedirectToAction("Index");
+            if (ModelState.IsValid)
+            {
+                _service.Delete(id);
+                return RedirectToAction("Index");
+
+            }
+            else
+            {
+                return View();
+            }
+           
         }
